Add XKeyStroke type (= java.awt.KeyStroke)
The XKeyStroke offers the same functionality as it's AWT counterpart,
but can be created idiomatically, with a list of the desired modifiers
as varargs.
It's now also easier to retrieve all the modifiers associated
with an XKeyStroke as list with the 'modifierObjects' property.

XKeyStroke additionally offers a 'stringSignature' property, to get the
string signature commonly seen in menus (e.g. Ctrl+Alt+X).

A new typealias for java.awt.Toolkit has been added in Toolkit.kt.