Implement XApplication
XApplication is the main entry point for all graphical applications
as well as a rudimentary, virtual base component.

XApplication itself is not a component, but has to be wrapped in a
container (XApplicationContainer). The configuration of the application
is incorporated into the container when the application is started.

An application can be started by calling the run() method, or executing
it in it's own Thread (since it's an instance of Runnable).

XApplicationTest.TestRun can be used to test run the XApplication API.

FatalException (added in this commit) is an exception used inside
XApplicationContainer to indicate fatal exceptions.