Change Vector.map() implementation return type
Return type has been changed from the implementing type to the
<>.Double implementation as standard.
Note that the map0() methods are not affected by this.

Since the default signature of map() requires the 'transform' function
to return a Double type anyway, using computing power to convert the
type and additionally loosing precision is unnecessary.

If the exact type (e.g. Point2.Int) is required, one can just call
map0(), given that they would need to have an instance of that exact
type to map on in the first place.