Fix bug in FusionComboBoxModel constructor
The loop copying the contents from the initial parameters to the
internal MutableList 'items' previously iterated over 'items.indices',
and because 'items' is empty prior to the copying process, indices
was equal to `0..0`, and nothing was copied.

The copying loop now iterates from 0 to the end of the collection
given as parameter (`0 until size`), fixing the bug.