Add Feature API and ClearingFeature
The Feature API allows extending components with additional
functionality without subclassing them, notably with text components.
A Feature can draw above it's connected component, and if it's an
instance of a listener interface, it is automatically added to it's
component as listener, enabling reactions to component events.

For now, the only component with the ability to host features is
XFeatureTextField, and automatic listener adding is restricted
to mouse listeners (mouse and motion).

ClearingFeature is intended for text components, where it adds an
additional button in the end of a text field, which clears the text
field when clicked.

The build script has been extended to allow running demo programs,
which helps testing visual and UI related code.
The ClearingFeature can be tested with the ClearingFeatureDemo class.