Change how demo running works
The previous :runDemos task had the demo class to execute set hard in
the configuration, making it unflexible and unsuitable for having more
than one demo class. Since not all demos have to be run at once,
the :runDemos task has been removed and replaced with a more flexible
:runDemo task.

The new task gets the demo class from the project's properties.
This way, the class can be specified dynamically by setting
the property 'demoClass'.

./gradlew runDemo -PdemoClass=<full demo class name>