Add TextPopupFeature
The feature shows a popup menu with actions to cut, copy, paste, delete
and select all text, enabling and disabling menu items contextually
(depending on selected text and the type of the component,
e.g. password fields will not have cut and copy option).
The menu does not show the key shortcuts deliberately, since they are
known to virtually every user, and would create unnecessary clutter
while conveying only unnecessary and superfluous information.

The model for this feature was the URL field popup menu in Firefox 65.