Fix state changing bug in ClearingFeature
The ClearingFeature would not change it's current state from HOVERED
to IDLE when the cursor exited the surrounding component at the end
of the feature's boundary, causing the feature to be drawn in
it's HOVERED state until the cursor entered the component again.

This has been fixed by adding a mouseExited() listener,
which is always changing the state to IDLE when the cursor exits
the component.