Add StyleBuilder factory class
The StyleBuilder works similar to the SubStyleBuilder, presenting
a mutable version of Style for the API consumer to edit.

The primary Style constructor has been changed
to take a Map<String, SubStyle>,
setting the 'subStyleMap' directly (only available internally)
The secondary constructor is the one destined for public use,
taking a Map<Selector, SubStyle>, transforming it and giving it
to the primary constructor.

This change makes the StyleBuilder more efficient, by providing
a constructor that skips the transformation step and enabling the
StyleBuilder to store a Map<String, SubStyle> internally.