Restructure Style constructors
The public constructor has been wrapped in the companion function
Style.of(Map<Selector, SubStyle>).

The direct setter constructor is now private, and the internal
constructor can now copy a given map automatically if it hasn't
been copied already, to avoid mutability issues and/or double-copying.

This constructor is now used by StyleBuilder.