Reorganize CSS processor
The processor is now organized along the flow of data,
represented by the Data interface.
Each instance of Data holds the data and the next step in the flow
(next() method).

The new process(String) method, called by loadCSS(String) iterates
down the flow until it gets a result off next() that is not an instance
of Data. If the result is the expected one, it is returned, otherwise
an Exception is thrown.

The different stages have been encapsulated in the Lexer and Parser
objects.
The Parser still stops at the step of breaking down the tokens into
blocks and declarations.