Add infix method combineWith() to SubStyle
This method can be used to combine two SubStyles that apply to the same
selector but come from different style sheets (or are defined multiple
times in the same style sheet).

The constructor of SubStyle is now similar to the constructor of Style,
and the immutable 'declarationMap' has been exposed internally.