Fix error in SubStyle.combineWith()
SubStyle.combineWith() was still using the old SubStyleBuilder.set(Any)
method, which did not exist anymore.

Since using a method similar to the old SubStyleBuilder.set() method
is the most straightforward way to perform a SubStyle combination
(avoiding casts and additional type checks), it is desirable to use
such a method.

The solution was to re-introduce this method, but this time with just
internal visibility, ensuring it can be misused.