Reorganize ValueParser code
The value parsing process has been split into two separate functions:
An internal parseValue() function that takes Tokens as parameter
and returns an Any.
It utilizes a private parseValue() function that accepts
a single token (String) and parses it to a value (Any).

This private function now contains most of the value parsing code,
which was previously the single-token branch of the internal
parseValue() function.